pyproject.toml
[tool.poetry]
name = "wildchord"
version = "0.1.0"
description = ""
authors = ["Mike Kurkin <mike@kurkin.net>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
dj-rest-auth = {extras = ["with_social"], version = "^2.2.5"}
djangorestframework-simplejwt = "^5.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
